+++ assets/js/app.js
@@ -1,7 +1,11 @@
 var submit = document.getElementById('transform_qr');
 var url = document.getElementById('url');
 var qr = new QRCode(
-  document.getElementById("qrcode")
+  document.getElementById("qrcode"), {
+    width: 128,
+    height: 128,
+    correctLevel : QRCode.CorrectLevel.H
+  }
 );
 
 submit.addEventListener('click', event => {